requirements.txt
pyyaml
jupyter
chardet
requests
pytest
pytest-cov
fabric3
tabulate
jupyter_contrib_nbextensions
sqlalchemy